Telephony: Notify ICC status immediately after registering
This change notifies the registrants for ICC_CHANGED, as soon as it
registers and thus it will receive the ICC_CHANGED event with latest ICC
status as soon as it registers. This is required because, sometimes(like
when switching radio tech) when the registrants register for the event,
they might not be notfied on the current ICC_STATUS as there is no
actual SIM/ICC Status change.

CRs-Fixed : 255301

Change-Id: I827e2a296cebafaac49dff41cf962356084c710e